msm: kgsl: Skip pagefault debugging if we can't get the lock
The nearby memory entries are nice to have in pagefault debugging
but we can hit some serious deadlocks if we try to wait for the
process mutex. Try to get the lock if we can but don't wait for it.
If we couldn't determine the process for some reason, just print
out a sorry message and move on with our lives.

Change-Id: Ic0dedbad0ee6ec9ca94e740f8de40b49cf7ed019
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Sunil Khatri <sunilkh@codeaurora.org>