soc: qcom: jtagv8: register for hotplug and cpu_pm notifiers
msm_jtag_save_state() and msm_jtag_restore_state() are called
directly from msm-pm driver. However msm-pm driver is going to be
deprecated when we move to PSCI framework. Add support in jtagv8
driver to rely on hotplug and cpu_pm callbacks for save and
restore functionality of Debug and ETM registers.

This commit also modifies the hotcpu callback in ETM driver to
avoid any race condition with hotcpu callback from Jtagv8 driver
that has been introduced in this commit.

Change-Id: I1261fdcb548b7a21b37090426efdbd77ea038333
Signed-off-by: Sarangdhar Joshi <spjoshi@codeaurora.org>
Signed-off-by: Xiaogang Cui <xiaogang@codeaurora.org>