mfd: wcd9335-regmap: Mark shared registers as volatile
Some of the codec registers can be shared with other processors and
hence can get updated without the knowledge of the driver. In such cases
the regmap cache is not in sync with the hardware register setting. Mark
these registers as volatile so that reads and writes bypass the cache.

Change-Id: I3c8e8e31e36f9f9a1aba0193d696d019437dfed2
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>