msm: mdss: add support to override dual DSI properties in kernel
In the current implementation, the DSI h/w configuration and DSI
PLL source are configured based on the MDSS DSI DT properties.
These can be changed dynamically based on the panel that is being
configured in LK. Add support to override these properties based on
the display command line string that is passed from LK to kernel.

Change-Id: I7ba5a2d16b3527537c91c79e7977a55d8b2e23c3
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>