emac: factor out all sgmii specific code
All code and resources which are specific to sgmii phy are moved
to spm_sgmii_v1 source file. The code includes register offsets,
fields definitions and parsing of device-tree. The resources
include two register spaces and a clock. The two register spaces
are actually part of the same one and thus consolidated in this patch.
The combined register space is updated in the devicetree.

Change-Id: Ic3e6f59a356e08173c3cd359f6605bc1c96d85a7
Signed-off-by: Gilad Avidov <gavidov@codeaurora.org>