ARM: dts: msm: correct variation in resonant frequency in LRA
Apply auto resonance correction algorithm on LRA actuators
if the frequency of the actuator drifts off from the resonant
frequency. Add DT property to enable support for the same.
CDP targets support both LRA and ERM. Hence, remove the change
which adds the change only for LRA actuator in 8976 CDP.

Change-Id: I59cf816d704b42cd2e2de28d87ba50e79b502259
Signed-off-by: Abinaya P <abinayap@codeaurora.org>