ASoC: wcd9335: Add support to enable vbat feature
Add support to enable vbat monitor feature on tasha codec. Vbat module
monitors the battery voltage and updates the gain on rx path if battery
volatage falls below certain threshold value.
This module can prevent brownout scenarios on the mobile devices when
the loads are heavy on the battery and it can also prolongs battery life.

Change-Id: I66bacf3b861618fc44f50a8e22ca3f2f0fc7763b
Signed-off-by: Venkata Narendra Kumar Gutta <vgutta@codeaurora.org>