ARM: dts: msm: increase the disconnect threshold for msm8956/msm8976
In some board designs, the transmit amplitude of the PHY is increased
to counter the attenuation resulted due to the additional components
between PHY and the connector. If the default disconnect threshold
is used, PHY may report false disconnect event. Hence increase the
disconnect threshold by 6% of the default value.

Change-Id: Ib1b2fe282b63683284974b6f958e5236d71cbba0
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>