regmap: Add regmap_multi_reg_write_bypassed function
Devices with more complex boot proceedures may occasionally apply the
register patch manual. regmap_multi_reg_write is a logical way to do so,
however the patch must be applied with cache bypass on, such that it
doesn't override any user settings. This patch adds a
regmap_multi_reg_write_bypassed function that applies a set of writes
with the bypass enabled.

Change-Id: I68755cd317b100f8c03dfdecace3e7b15068e066
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Santosh Mardi <gsantosh@codeaurora.org>