drivers: Compile the clk source tree sooner
Most of the drivers requires the appropriate clocks propagated in the
clock tree when probing the device. By changing the order in the Makefile,
the clk source tree can be compiled much earlier so that the init function
to probe the clock can be appear at the beginning of arch_initcall table.

Change-Id: I1786530b318778c18357d9b30ccb639c0c8e533e
Acked-by: Jie Luo <jluo@qti.qualcomm.com>
Signed-off-by: Kranthikumar Kurapati <kkurap@codeaurora.org>