msm: mdss: refine logic for configuration of DSI clock sources
The DSI clocks need to sourced out of the correct DSI PLL based
on the h/w configuration. The current assumption is that the
clk_set_parent operation on DSI branch clocks is necessary for
targets which support 2 DSI PLLs. However, there is a possibility
that the DSI clock source needs to be set even for single DSI
targets if the clock driver implements multi-parent based
approach for DSI RCG clock structures. Add change to take care
of this.

Change-Id: I074e1db7d1f5a34fabd18ef5dcc586b8e0d0e200
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>