msm: mdss: add support to swap the frame content on interfaces
For some split DSI panels, there is a requirement to swap the
content sent to both the DDICs due to some constraints at the
panel end. Add support to achieve this in cases when Pingpong
Split is enabled at MDP side. In cases when LM split is used,
it would be sufficient if the Mixer Swap variable is set to
true. Rest of the support is already available for this case.

Change-Id: I47195b4dd7ce4fa775106fbdfef62b3bb26fdce1
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>