drivers: Compile the clk source tree earlier
Most of the drivers require the appropriate clocks propagated in the
clock tree when probing the device. By changing the order in the Makefile,
the clk source tree can be compiled much earlier so that the init function
to probe the clock can be appear at the beginning of arch_initcall table.

Change-Id: Ide35d5ea06b381368d7929dfb12e7c0e5382024f
Acked-by: Jie Luo <jluo@qti.qualcomm.com>
Signed-off-by: Kranthikumar Kurapati <kkurap@codeaurora.org>