msm: ipa: discard holb mitigation on Q6 pipes in ssr
When modem is shutdown, Q6 pipes are reset through uc.
As pipes are reset, holb mitigation logic kicks in
and ep_delay is set on all consumer pipes. At this point
if USB data transfer is going on and if USB is disconnected
IPA will not consume packets from USB and this results in
a forced crash. Make changes to discard holb monitoring on
Q6 pipes when modem is shutdown and enable back when QMI
handshake is completed. This will ensure holb monitoring
will not kick in when pipes are resetted.

Change-Id: Iaed23bb61b43eceef36c1fc5d264825d55e7268f
Acked-by: Chaitanya Pratapa <cpratapa@qti.qualcomm.com>
Signed-off-by: Ravinder Konka <rkonka@codeaurora.org>