msm: thermal: Update the inline definition for BWLM API
KTM header has inline declaration for the bandwidth limiting
feature APIs, when thermal is not defined.
This will have compilation errors on targets with no KTM driver,
as the function is not defined, but just declared in the header.

Update the header to have an inline definition for the BWLM API
to return ENOSYS.

Change-Id: I1864c409c3941cd4a6a1abaaf0385cedae58da92
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>